move swagger middleware setup into separate file
